+++ selfdrive/boardd/panda.cc
@@ -2,9 +2,32 @@
 #include <cassert>
 #include <iostream>
 
+#include <unistd.h>
+
 #include "common/swaglog.h"
+#include "common/gpio.h"
 
 #include "panda.h"
+
+void panda_set_power(bool power){
+#ifdef QCOM2
+  int err = 0;
+  err += gpio_init(GPIO_STM_RST_N, true);
+  err += gpio_init(GPIO_STM_BOOT0, true);
+  err += gpio_init(GPIO_HUB_RST_N, true);
+
+  err += gpio_set(GPIO_STM_RST_N, false);
+
+  // TODO: set hub somewhere else
+  err += gpio_set(GPIO_HUB_RST_N, true);
+  err += gpio_set(GPIO_STM_BOOT0, false);
+
+  usleep(100*1000); // 100 ms
+
+  err += gpio_set(GPIO_STM_RST_N, power);
+  assert(err == 0);
+#endif
+}
 
 Panda::Panda(){
   int err;
@@ -25,6 +48,10 @@ Panda::Panda(){
   dev_handle = libusb_open_device_with_vid_pid(ctx, 0xbbaa, 0xddcc);
   if (dev_handle == NULL) { goto fail; }
 
+  if (libusb_kernel_driver_active(dev_handle, 0) == 1) {
+    libusb_detach_kernel_driver(dev_handle, 0);
+  }
+
   err = libusb_set_configuration(dev_handle, 1);
   if (err != 0) { goto fail; }
 
@@ -35,8 +62,10 @@ Panda::Panda(){
   is_pigeon =
     (hw_type == cereal::HealthData::HwType::GREY_PANDA) ||
     (hw_type == cereal::HealthData::HwType::BLACK_PANDA) ||
-    (hw_type == cereal::HealthData::HwType::UNO);
-  has_rtc = (hw_type == cereal::HealthData::HwType::UNO);
+    (hw_type == cereal::HealthData::HwType::UNO) || 
+    (hw_type == cereal::HealthData::HwType::DOS);
+  has_rtc = (hw_type == cereal::HealthData::HwType::UNO) ||
+    (hw_type == cereal::HealthData::HwType::DOS);
 
   return;