+++ selfdrive/boardd/panda.h
@@ -34,6 +34,9 @@ struct __attribute__((packed)) health_t {
   uint8_t power_save_enabled;
 };
 
+
+void panda_set_power(bool power);
+
 class Panda {
  private:
   libusb_context *ctx = NULL;