+++ selfdrive/modeld/models/driving.h
@@ -9,11 +9,13 @@
 
 #include "common/mat.h"
 #include "common/util.h"
+#include "common/modeldata.h"
 
 #include "commonmodel.h"
 #include "runners/run.h"
 
 #include <czmq.h>
+#include <memory>
 #include "messaging.hpp"
 
 #define MODEL_WIDTH 512
@@ -21,13 +23,8 @@
 #define MODEL_FRAME_SIZE MODEL_WIDTH * MODEL_HEIGHT * 3 / 2
 #define MODEL_NAME "supercombo_dlc"
 
-#define MODEL_PATH_DISTANCE 192
-#define POLYFIT_DEGREE 4
-#define SPEED_PERCENTILES 10
 #define DESIRE_LEN 8
 #define TRAFFIC_CONVENTION_LEN 2
-#define DESIRE_PRED_SIZE 32
-#define OTHER_META_SIZE 4
 #define LEAD_MDN_N 5 // probs for 5 groups
 #define MDN_VALS 4 // output xyva for each lead group
 #define SELECTION 3 //output 3 group (lead now, in 2s and 6s)
@@ -58,11 +55,11 @@ typedef struct ModelState {
   float *input_frames;
   RunModel *m;
 #ifdef DESIRE
-  float *prev_desire;
-  float *pulse_desire;
+  std::unique_ptr<float[]> prev_desire;
+  std::unique_ptr<float[]> pulse_desire;
 #endif
 #ifdef TRAFFIC_CONVENTION
-  float *traffic_convention;
+  std::unique_ptr<float[]> traffic_convention;
 #endif
 } ModelState;