+++ selfdrive/test/test_car_models.py
@@ -172,6 +172,10 @@ routes = {
     'carFingerprint': HYUNDAI.HYUNDAI_GENESIS,
     'enableCamera': True,
   },
+  "70c5bec28ec8e345|2020-08-08--12-22-23": {
+    'carFingerprint': HYUNDAI.GENESIS_G70,
+    'enableCamera': True,
+  },
   "38bfd238edecbcd7|2018-08-22--09-45-44": {
     'carFingerprint': HYUNDAI.SANTA_FE,
     'enableCamera': False,
@@ -206,6 +210,10 @@ routes = {
   },
   "2c5cf2dd6102e5da|2020-06-26--16-00-08": {
     'carFingerprint': HYUNDAI.IONIQ,
+    'enableCamera': True,
+  },
+  "22d955b2cd499c22|2020-08-10--19-58-21": {
+    'carFingerprint': HYUNDAI.KONA,
     'enableCamera': True,
   },
   "5dddcbca6eb66c62|2020-07-26--13-24-19": {
@@ -389,6 +397,26 @@ routes = {
     'carFingerprint': SUBARU.IMPREZA,
     'enableCamera': True,
   },
+  # Dashcam
+  "95441c38ae8c130e|2020-06-08--12-10-17": {
+    'carFingerprint': SUBARU.FORESTER_PREGLOBAL,
+    'enableCamera': True,
+  },
+  # Dashcam
+  "df5ca7660000fba8|2020-06-16--17-37-19": {
+    'carFingerprint': SUBARU.LEGACY_PREGLOBAL,
+    'enableCamera': True,
+  },
+  # Dashcam
+  "5ab784f361e19b78|2020-06-08--16-30-41": {
+    'carFingerprint': SUBARU.OUTBACK_PREGLOBAL,
+    'enableCamera': True,
+  },
+  # Dashcam
+  "e19eb5d5353b1ac1|2020-08-09--14-37-56": {
+    'carFingerprint': SUBARU.OUTBACK_PREGLOBAL_2018,
+    'enableCamera': True,
+  },
   "fbbfa6af821552b9|2020-03-03--08-09-43": {
     'carFingerprint': NISSAN.XTRAIL,
     'enableCamera': True,
@@ -437,7 +465,6 @@ non_tested_cars = [
   HYUNDAI.GENESIS_G90,
   HYUNDAI.KIA_FORTE,
   HYUNDAI.KIA_OPTIMA_H,
-  HYUNDAI.KONA,
   HYUNDAI.KONA_EV,
   TOYOTA.CAMRYH,
   TOYOTA.CHR,