+++ tools/sim/launch_openpilot.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+export PASSIVE="0"
+export NOBOARD="1"
+
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
+cd ../../selfdrive && ./manager.py